chore: move interfaces to new files
