+++ react-components/src/components/CacheProvider/CacheProvider.ts
@@ -1,11 +1,11 @@
 import { type CdfCaches } from '../../architecture/base/renderTarget/CdfCaches';
 import { useRenderTarget } from '../RevealCanvas';
 import { type ClassicCadAssetMappingCache } from './cad/ClassicCadAssetMappingCache';
-import { type FdmCadNodeCache } from './cad/FdmCadNodeCache';
 import { type PointCloudAnnotationCache } from './PointCloudAnnotationCache';
 import { type Image360AnnotationCache } from './Image360AnnotationCache';
 import { type Fdm3dDataProvider } from '../../data-providers/Fdm3dDataProvider';
 import { type CadInstanceMappingsCache } from './cad/CadInstanceMappingsCache';
+import { FdmCadNodeCache } from './cad/FdmCadNodeCache';
 
 const useCacheObject = (): CdfCaches => {
   const revealRenderTarget = useRenderTarget();
