+++ react-components/src/hooks/cad/useAssetMappingForTreeIndex.ts
@@ -1,9 +1,9 @@
 import { useQuery, type UseQueryResult } from '@tanstack/react-query';
 import { fetchAncestorNodesForTreeIndex } from '../../components/CacheProvider/requests';
 import { type AnyIntersection } from '@cognite/reveal';
-import { type ClassicCadNodeAssetMappingResult } from '../../components/CacheProvider/cad/ClassicCadAssetMappingCache';
 import { useSDK } from '../../components/RevealCanvas/SDKProvider';
 import { useClassicCadAssetMappingCache } from '../../components/CacheProvider/CacheProvider';
+import { type ClassicCadNodeAssetMappingResult } from '../../components/CacheProvider/cad/ClassicCadAssetMappingCache';
 
 export const useAssetMappingForTreeIndex = (
   intersection: AnyIntersection | undefined
